egx_streaming_pipeline/producer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
kafka-python==2.0.2
